refactor generator utilities into shared modules
